chore(zui): rename transformer dir to effect
