+++ packages/zui/src/z/types/index.ts
@@ -30,7 +30,7 @@ export * from './ref'
 export * from './set'
 export * from './string'
 export * from './symbol'
-export * from './transformer'
+export * from './effects'
 export * from './tuple'
 export * from './undefined'
 export * from './union'
